Update AEAudioUnitFilter.m
Before attempting to set the controller's stream format as the input format of the audio unit, we ask the unit for it's input property. If this is different than the controller's format, we create a converter node. Otherwise, we proceed with attempting to set the controllers format on the node.

This allows something like the 3D mixer unit to behave properly, since although it requires mono input to function correctly, it does not actually complain when a stereo format is applied to its input.