Changed Audiobus input support to use Audiobus's live mode for simplicity
Previously, AEAudioController's audiobusInputPort was used to receive audio from Audiobus via the "lossless" stream mode. This ensured audio received was identical to audio sent from the Audiobus source, but at the cost of increased latency and unpredictable audio receive intervals, which increased the complexity of receiving apps due to the necessity of dealing with buffer under runs by using a separate "live" stream for monitoring (using AEPlaythroughBuffer or manually).

This change causes AEAudioController to only use Audiobus' "live" mode for input, removing the necessity of dealing with the particulars of "lossless" mode with little tradeoff, as live mode yields excellent results.