Replaced "avoidMeasurementModeForBuiltInMic" feature with input boost
Prior to this commit, by default, Measurement Mode was not used with the built-in mic due to an extremely low audio level. Given that, as of the iPad Air 2, input latency is extremely high without Measurement Mode, it was important to be able to use this mode with the device mic. So, this commit now causes TAAE to automatically boost the input level by default (by an empirically determined linear factor of 4, or 12dB), when the built-in mic is used with Measurement Mode.