Added completion blocks to most add/remove methods
TAAE transitioned a few years back to async operations, but was missing completion blocks so there was no guarantee operations were completed by the time the methods returned, causing problems with operations that depended on the last step (like setting an Audiobus sender port for an added channel). This commit adds completion blocks to address the issue.