Use applicationDidBecomeActive rather than applicationWillEnterForeground for restarting the audio session
This partially works around an error in iOS 12.4 which causes AudioOutputUnitStart to fail with AVAudioSessionErrorCodeCannotStartRecording if the app is hosted within IAA and launched into the background, and persists even when the app enters the foreground if the IO audio unit is started again during UIApplicationWillEnterForegroundNotification. If the IO audio unit is started in UIApplicationDidBecomeActiveNotification, however, the operation succeeds.